# internal/api/server/server.gen.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.1-0.20260318123712-00a90b7a03f4 DO NOT EDIT.
package server

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	. "github.com/dcm-project/k8s-storage-service-provider/api/v1alpha1"
	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List Volumes
	// (GET /api/v1alpha1/volumes)
	ListVolumes(w http.ResponseWriter, r *http.Request, params ListVolumesParams)
	// Create Volume
	// (POST /api/v1alpha1/volumes)
	CreateVolume(w http.ResponseWriter, r *http.Request, params CreateVolumeParams)
	// Health Check
	// (GET /api/v1alpha1/volumes/health)
	GetHealth(w http.ResponseWriter, r *http.Request)
	// Delete Volume
	// (DELETE /api/v1alpha1/volumes/{volume_id})
	DeleteVolume(w http.ResponseWriter, r *http.Request, volumeId VolumeIdPath)
	// Get Volume
	// (GET /api/v1alpha1/volumes/{volume_id})
	GetVolume(w http.ResponseWriter, r *http.Request, volumeId VolumeIdPath)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// List Volumes
// (GET /api/v1alpha1/volumes)
func (_ Unimplemented) ListVolumes(w http.ResponseWriter, r *http.Request, params ListVolumesParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create Volume
// (POST /api/v1alpha1/volumes)
func (_ Unimplemented) CreateVolume(w http.ResponseWriter, r *http.Request, params CreateVolumeParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Health Check
// (GET /api/v1alpha1/volumes/health)
func (_ Unimplemented) GetHealth(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Delete Volume
// (DELETE /api/v1alpha1/volumes/{volume_id})
func (_ Unimplemented) DeleteVolume(w http.ResponseWriter, r *http.Request, volumeId VolumeIdPath) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get Volume
// (GET /api/v1alpha1/volumes/{volume_id})
func (_ Unimplemented) GetVolume(w http.ResponseWriter, r *http.Request, volumeId VolumeIdPath) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListVolumes operation middleware
func (siw *ServerInterfaceWrapper) ListVolumes(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListVolumesParams

	// ------------- Optional query parameter "max_page_size" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "max_page_size", r.URL.Query(), &params.MaxPageSize, runtime.BindQueryParameterOptions{Type: "integer", Format: "int32"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "max_page_size", Err: err})
		return
	}

	// ------------- Optional query parameter "page_token" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "page_token", r.URL.Query(), &params.PageToken, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page_token", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListVolumes(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateVolume operation middleware
func (siw *ServerInterfaceWrapper) CreateVolume(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params CreateVolumeParams

	// ------------- Optional query parameter "id" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "id", r.URL.Query(), &params.Id, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateVolume(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteVolume operation middleware
func (siw *ServerInterfaceWrapper) DeleteVolume(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "volume_id" -------------
	var volumeId VolumeIdPath

	err = runtime.BindStyledParameterWithOptions("simple", "volume_id", chi.URLParam(r, "volume_id"), &volumeId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "volume_id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteVolume(w, r, volumeId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetVolume operation middleware
func (siw *ServerInterfaceWrapper) GetVolume(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "volume_id" -------------
	var volumeId VolumeIdPath

	err = runtime.BindStyledParameterWithOptions("simple", "volume_id", chi.URLParam(r, "volume_id"), &volumeId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "volume_id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetVolume(w, r, volumeId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1alpha1/volumes", wrapper.ListVolumes)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/v1alpha1/volumes", wrapper.CreateVolume)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1alpha1/volumes/health", wrapper.GetHealth)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/api/v1alpha1/volumes/{volume_id}", wrapper.DeleteVolume)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1alpha1/volumes/{volume_id}", wrapper.GetVolume)
	})

	return r
}

type ListVolumesRequestObject struct {
	Params ListVolumesParams
}

type ListVolumesResponseObject interface {
	VisitListVolumesResponse(w http.ResponseWriter) error
}

type ListVolumes200JSONResponse VolumeList

func (response ListVolumes200JSONResponse) VisitListVolumesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type ListVolumes400ApplicationProblemPlusJSONResponse Error

func (response ListVolumes400ApplicationProblemPlusJSONResponse) VisitListVolumesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type ListVolumes500ApplicationProblemPlusJSONResponse Error

func (response ListVolumes500ApplicationProblemPlusJSONResponse) VisitListVolumesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type CreateVolumeRequestObject struct {
	Params CreateVolumeParams
	Body   *CreateVolumeJSONRequestBody
}

type CreateVolumeResponseObject interface {
	VisitCreateVolumeResponse(w http.ResponseWriter) error
}

type CreateVolume201JSONResponse Volume

func (response CreateVolume201JSONResponse) VisitCreateVolumeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type CreateVolume400ApplicationProblemPlusJSONResponse Error

func (response CreateVolume400ApplicationProblemPlusJSONResponse) VisitCreateVolumeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type CreateVolume409ApplicationProblemPlusJSONResponse Error

func (response CreateVolume409ApplicationProblemPlusJSONResponse) VisitCreateVolumeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

type CreateVolume422ApplicationProblemPlusJSONResponse Error

func (response CreateVolume422ApplicationProblemPlusJSONResponse) VisitCreateVolumeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(422)
	_, err := buf.WriteTo(w)
	return err
}

type CreateVolume500ApplicationProblemPlusJSONResponse Error

func (response CreateVolume500ApplicationProblemPlusJSONResponse) VisitCreateVolumeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetHealthRequestObject struct {
}

type GetHealthResponseObject interface {
	VisitGetHealthResponse(w http.ResponseWriter) error
}

type GetHealth200JSONResponse Health

func (response GetHealth200JSONResponse) VisitGetHealthResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type DeleteVolumeRequestObject struct {
	VolumeId VolumeIdPath `json:"volume_id"`
}

type DeleteVolumeResponseObject interface {
	VisitDeleteVolumeResponse(w http.ResponseWriter) error
}

type DeleteVolume204Response struct {
}

func (response DeleteVolume204Response) VisitDeleteVolumeResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type DeleteVolume404ApplicationProblemPlusJSONResponse Error

func (response DeleteVolume404ApplicationProblemPlusJSONResponse) VisitDeleteVolumeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type DeleteVolume500ApplicationProblemPlusJSONResponse Error

func (response DeleteVolume500ApplicationProblemPlusJSONResponse) VisitDeleteVolumeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetVolumeRequestObject struct {
	VolumeId VolumeIdPath `json:"volume_id"`
}

type GetVolumeResponseObject interface {
	VisitGetVolumeResponse(w http.ResponseWriter) error
}

type GetVolume200JSONResponse Volume

func (response GetVolume200JSONResponse) VisitGetVolumeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetVolume404ApplicationProblemPlusJSONResponse Error

func (response GetVolume404ApplicationProblemPlusJSONResponse) VisitGetVolumeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type GetVolume500ApplicationProblemPlusJSONResponse Error

func (response GetVolume500ApplicationProblemPlusJSONResponse) VisitGetVolumeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// List Volumes
	// (GET /api/v1alpha1/volumes)
	ListVolumes(ctx context.Context, request ListVolumesRequestObject) (ListVolumesResponseObject, error)
	// Create Volume
	// (POST /api/v1alpha1/volumes)
	CreateVolume(ctx context.Context, request CreateVolumeRequestObject) (CreateVolumeResponseObject, error)
	// Health Check
	// (GET /api/v1alpha1/volumes/health)
	GetHealth(ctx context.Context, request GetHealthRequestObject) (GetHealthResponseObject, error)
	// Delete Volume
	// (DELETE /api/v1alpha1/volumes/{volume_id})
	DeleteVolume(ctx context.Context, request DeleteVolumeRequestObject) (DeleteVolumeResponseObject, error)
	// Get Volume
	// (GET /api/v1alpha1/volumes/{volume_id})
	GetVolume(ctx context.Context, request GetVolumeRequestObject) (GetVolumeResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// ListVolumes operation middleware
func (sh *strictHandler) ListVolumes(w http.ResponseWriter, r *http.Request, params ListVolumesParams) {
	var request ListVolumesRequestObject

	request.Params = params

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ListVolumes(ctx, request.(ListVolumesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListVolumes")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ListVolumesResponseObject); ok {
		if err := validResponse.VisitListVolumesResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CreateVolume operation middleware
func (sh *strictHandler) CreateVolume(w http.ResponseWriter, r *http.Request, params CreateVolumeParams) {
	var request CreateVolumeRequestObject

	request.Params = params

	var body CreateVolumeJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreateVolume(ctx, request.(CreateVolumeRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateVolume")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreateVolumeResponseObject); ok {
		if err := validResponse.VisitCreateVolumeResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetHealth operation middleware
func (sh *strictHandler) GetHealth(w http.ResponseWriter, r *http.Request) {
	var request GetHealthRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetHealth(ctx, request.(GetHealthRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetHealth")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetHealthResponseObject); ok {
		if err := validResponse.VisitGetHealthResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// DeleteVolume operation middleware
func (sh *strictHandler) DeleteVolume(w http.ResponseWriter, r *http.Request, volumeId VolumeIdPath) {
	var request DeleteVolumeRequestObject

	request.VolumeId = volumeId

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteVolume(ctx, request.(DeleteVolumeRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteVolume")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(DeleteVolumeResponseObject); ok {
		if err := validResponse.VisitDeleteVolumeResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetVolume operation middleware
func (sh *strictHandler) GetVolume(w http.ResponseWriter, r *http.Request, volumeId VolumeIdPath) {
	var request GetVolumeRequestObject

	request.VolumeId = volumeId

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetVolume(ctx, request.(GetVolumeRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetVolume")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetVolumeResponseObject); ok {
		if err := validResponse.VisitGetVolumeResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}